refactor: split cmd/roborev/main.go into focused files
Extract ~3,500 lines from main.go into 19 focused files following the
established pattern (same package, access globals directly, xxxCmd()
returns *cobra.Command).

main.go is reduced to ~60 lines: global vars, main() with root command
setup, and AddCommand calls.

New files:
- env.go: git environment filtering, binary detection helpers
- daemon_lifecycle.go: test seams, daemon start/stop/restart
- daemon_client.go: HTTP client helpers (waitForJob, showReview, etc.)
- helpers.go: exitError, shortRef, autoInstallHooks, etc.
- update.go: update command + PID-tracking restart logic
- review.go: review command + runLocalReview
- daemon_cmd.go: daemon subcommand group + daemonRunCmd
- init_cmd.go: init command
- hook.go: install-hook, uninstall-hook
- wait.go, status.go, list.go, show.go, comment.go, stream.go
- skills.go, sync.go, check_agents.go, version.go

No behavioral changes. Tests unmodified and passing.
