+++ cmd/roborev/version.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/roborev-dev/roborev/internal/version"
+	"github.com/spf13/cobra"
+)
+
+func versionCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "version",
+		Short: "Show roborev version",
+		Run: func(cmd *cobra.Command, args []string) {
+			fmt.Printf("roborev %s\n", version.Version)
+		},
+	}
+}
